implement avatar fetching for buddies
this adds fields in the SlackUser for avatar_hash and avatar_url that
are populated while fetching the users list, after the buddy is created
in slack_im_set() we make a call to a new function slack_avatar_update()
to fetch the avatar. Letting this happen after we actually have the
PurpleBuddy allows caching and not downloading avatars for users that
are not on our list.